requirements.txt
gradio==3.5.0
torch==2.0.1
transformers==4.32.0
Pillow==9.5.0
